Fix share submit when exceed upstream capacity
When downstream send a share but we exceed the upstream capacity (70
shares per minute) we are not going to send the share upstream. In that
case before this commit we were sending downstream a share submission
error even when the share is valid. We fix it and check the share target
and send an error ONLY when the share is invalid.